+++ src/errors.ts
@@ -9,4 +9,6 @@ export class MoreThanOneDistinctOrAllError extends Error{}
 
 export class MoreThanOneWhereStepError extends Error{}
 
+export class InvalidConditionError extends Error{}
+
 export class DeleteWithoutConditionError extends Error{}