+++ CHANGELOG.md
@@ -12,6 +12,13 @@
 - Function selectAsteriskFrom() now can take TableAliases not just Tables
 - Rename `WhereStep` class to `SelectWhereStep`
 
+## [0.12.1]  2022-07-02
+### Changed
+- Update README.md: fix code and add railroad diagram to it and add its svg file to publish
+- Check validity of Condition
+- Add more types to function eq() in Expression
+- Add function ne() to Expression
+
 ## [0.12.0]  2022-06-20
 ### Added
 - Support Date Column, this is a new feature to support Date and timestamp with and without timezone columns.